+++ CHANGELOG.md
@@ -8,3 +8,6 @@
 - Phoneme builders for accent phoneme sets
 - General American English phoneme set
 
+### Changed
+- Phoneme symbol is char (not string)
+- Phoneme implements Copy